requirements.txt
google-genai
edge-tts
pydub